Fix: cleanup DataProvider: private methods and removed unused
