+++ java/jscope/src/main/java/mds/provider/MdsDataProviderUdt.java
@@ -1,9 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 package mds.provider;
 
 import mds.connection.MdsConnection;
@@ -15,9 +9,7 @@ import mds.provider.mds.MdsConnectionUdt;
  */
 public class MdsDataProviderUdt extends MdsDataProvider
 {
-     protected MdsConnection getConnection()
-    {
-	return new MdsConnectionUdt();
-    }
-
+	@Override
+	protected MdsConnection getConnection()
+	{ return new MdsConnectionUdt(); }
 }
